+++ TutorialApp.Datalayer/Entities/Quiz.cs
@@ -10,7 +10,7 @@ namespace TutorialApp.Datalayer.Entities
     public class Quiz
     {
         [Key]
-        public long QuizId { get; set; }
+        public int QuizId { get; set; }
         [Display(Name = "title of quiz")]
         [Required(ErrorMessage = "The {0} must not be empty.")]
         [MaxLength(100, ErrorMessage = "The {0} can not be more than {1} characters.")]
@@ -23,7 +23,7 @@ namespace TutorialApp.Datalayer.Entities
         //Relation Question
         public virtual ICollection<Question> Questions { get; set; }
         //Foreign key with Course
-        public virtual long CourseId { get; set; }
+        public virtual int CourseId { get; set; }
         public virtual Course Course { get; set; }
         #endregion
     }
